src/conftest: Add S() fixture for the symtab of M() fixture's machine
This gets rid of the `S = M.symtab` repeated in many tests.

It relies on a fixture being called only once, even if its result must be
passed to other fixtures as well as the test function. I haven't seen
documentation of this behaviour, but it works and it does seem to make
sense given that pytest has careful scope control for each fixture.