testmc.mc6800: Moved from mc6800sim, and MC6800 renamed to Machine
Under `testmc` we have submodules defining the test machine, `m6502` for
the wrapper around py65 and `mc6800` for the 6800 simulator-in-progress.

The test machine itself is now always named `….machine.Machine` because
that consistency seems more clearly to show that they share a common
interface. (We can always use import renaming if we need to have more than
one Machine available within a module.) This will also make it easier to
write generic test framework code to do thing on whatever particular
Machine a test module has imported.

This means that `testmc` is now going to include the entire 6800 simulator,
but given that it's fairly heavily dependent on the testmc code and is
designed as a simulator for unit tests, rather than a generic simulator,
there doesn't seem to be much point in trying set it up to become a
separate package.