src/simple-a68: Create mc6800.Machine and load object code into it
This changes our test framework to parametrize the Machine type; rather
than always using a testmc.m6502.Machine, we now use whatever Machine
has been imported by the caller, so we use whatever Machine the tests
expect. Cleverly, this requires no changes to the tests themselves.

The M, S and R fixtures really should be moved to the testmc module
at some point, since they're generic to anything using it.

We test that the correct data are being loaded into a 6800 Machine and that
we can step an LDAA instruction (the only one that's implemented); further
tests for call() will have to wait until we have an RTS instruction and
call() is implemented.