testmc.generic.machine load_memimage(): Pull up from testmc.m6502
This routine is mostly generic and so pulls up with only one change: we now
use setregs() to set the program counter `pc` to the entrypoint of the
file.

The load() function has some moderately complex logic but not unit tests
for it; it's currenly tested by the code under `src/` that is a client of
`testmc`. This is probably true of other code as well and bringing proper
test coverage into `testmc` itself may be a bit of work, so just add an
xfail'd test for the moment and we'll deal with this later. (It will be
necessary to do this when we extract `testmc` into its own project.)