testmc.mc6800: Implement RTS
We now offer the option to pass a machine to runop() in the `opcodes` tests
so that we can configure specific states before executing the opcode.

We also add a popword() utility routine to `opimpl` to pop a word off the
stack and properly handle wrapping. We have to pop a word with a pair of
byte accesses because the stack pointer may wrap across the two bytes.
(That 6800 handles it this way is made clear in the Programming Reference
Manual, which describes RTS as incrementing the SP and setting PCH and PCL
in two steps.)