testmc.generic.machine: Make is_little_endian a property
This is a bit cleaner and shorter to define as a property rather than a
method.

According to the Python standard library docs, @abstractproperty is
deprecated and one should use @abstractmethod followed by @property.
However, that alone produces a read-only property which prevents subclasses
authors from writing just `is_little_endian = True`. So rather than add the
extra verbosity of defining an additional `@is_little_endian.setter
@abstractmethod`, just use @abstractproperty anyway. (We may need to review
this decision if we start checking for deprecation warnings.)