testmc.generic.machine stepto(): Pull up from m6502
It turns out that Machine.regs is quite slow, which is not particularly
surprising given that it does quite a lot of work in the constructor to
check values. stepto() needs to fetch the program counter value through
this in order to look up the next byte, and doing it via Machine.regs
slowed the function massively (by an order of magnitude or more). So we've
added a _getpc() function and force subclasses to give us a faster way of
fetching the PC.

It's possible this could be applied to other calls going through the
layers, such as byte(), but it doesn't seem worthwhile at the moment.

As is usual, we leave the m6502 tests in place to help confirm that it
still works correctly; these can be trimmed down to a smoke test in a later
commit.