testmc.m6502.machine call(): Extract _{JSR,RTS}_instructions attributes
Rather than hardcode the opcodes for JSR and RTS used by call() we now
fetch them from attributes. This will allow us to vary the different sets
of opcodes used by different CPUs when we later pull call() up to the
superclass. These attributes must be sets or similar objects supporting the
`|` operator for set union.

Though each set contains only a single opcode on the 6502, we make these
attributes sets because other CPUs have multiple opcodes for calling and
returning from subroutines (e.g., BSR and two addressing modes of JSR on
6800). However, this does not yet handle conditional calls and returns, as
on the Intel 8080; we'll later have to figure out what to do in those
cases.

We also extract an attribute for the default "abort" obcodes, i.e., those
which cause call() to stop execution with an Abort exception. See the code
comments for further details on that.