testmc..machine stepto(): stopon parameter reworking
stepto()'s `stopon` parameter is now a named parameter only, optional, and
must be a Container.

We make it named and optional in anticipation of adding a `stopat`
parameter, allowing us to stop at location as well as or instead of an
opcode. It now defaults to an empty set.

We now disallow scalar values, rather than trying to detect them and
convert them to a singleton set. We were starting to think about allowing
things like sets of ranges and realized that that way lies insanity; best
to have the caller create and/or combine whatever set-like objects he needs
that can provide their own membership tests as efficient (or not) as
necessary. (This is thinking forward to `stopat`, where one might want
quite large ranges of addresses in which to stop.)

As part of this cleanup we also:
- Simplify the stepto() code.
- Improve the stepto() docstring.
- Rename test_stepto_empty_stopon_and_otherargs() to test_stepto_maxsteps()
  and move it to the start of the step() tests.