testmc..registers: Valued/non-valued terminology and .valued()/.all() methods
Improve the terminology we use for the registers and flags in Registers()
objects: each is _valued_ if it has a non-None value, or _non-valued_ if
its value is None, meaning "don't care."

Rename the dict() method to valued(), to make it clear that it's not
returning dictionary entries for non-valued registers and flags, and add an
all() method to return a dictionary that includes the non-valued registers
and flags.

Probably valued() should be refactored to use all(); the code is very
similar between the two.