testmc.mc6800.opcodes: Improve before/after Register checks in runop() tests
Summary:
1. Checked registers not set at start are automatically set to a value
   other than the checked one, to confirm it was actually changed.
2. After the op, assert that the machine has no unvalued registers.
3. Improve the printing of expected and actual register values.

*** 1. Setting registers/flags to unexpected values

Before running the opcode, the runop() function now changes any registers
and flags that are tested (excepting PC and SP) to something other than the
expected value if they were not preset by the caller.

This finds cases where the returned value is always the flag default value
but it wasn't explicitly set. For example, LDAA must clear the overflow
flag, but previously if you removed the line that did that the introduced
bug would not be detected because it was always clear when entering that
function.

This doesn't entirely replace developers setting up correct test conditions
to detect this sort of thing, but it does partially automate it.

*** 2. Assert machine has no unvalued registers/flags

While Registers objects used for comparisons can have don't-care values,
obviously the Machine itself never can. After running the opcode, scan all
registers and flags in the Machine to ensure that they are valued. If
they're not, the opcode implementation has probably done something terribly
wrong, like forgetting a `return` statement in a function (this really
happened).

Possibly we should do this check before we run the opcode, too, but it
doesn't really seem worth the effort. And who knows, someone might even
want to build tests with the machine having non-valued registers or flags
in the initial state.

*** 3. Improve printed output

As well as the expected valued registers (and flags), we now also print all
registers (including the values we expect not to change) and the actual
values, on three separate lines to stdout. This makes it easier to see, on
failure, what went wrong.

Also we extract the regs from the Machine outside of the assert so that
pytest doesn't start digging down to print additional lines deconstructing
the source, as:

    where 6800 pc=0010 a=A0 b=01 x=0001 sp=BFFF --HInZVC = <testmc.mc6800.machine.Machine object at 0x7fab869ea4a8>.regs

Probably we should be fixing the way that the Machine object is repr()'d,
however.