testmc.generic.iomem: Memory with mock device I/O
This is a memory implementation backed by a bytearray, but adding the
ability to have reads from and writes to arbitrary locations call Python
functions. (This allows you to simulate I/O.)

It also adds some additional safety features to help detect bad use of
memory, including disallowing length changes and raising `IndexError` for
negative indexes and slices extending beyond the end of the sequence.

It does not currently support supplying sequences or ranges for I/O
addresses; each address must be set individually. This should be fixed.