testmc.mc6800: Implement LSRA
This is probably, but not certainly, has correct V flag behaviour. It's
confusing because it appears that the V flag is meaningless for LSR and
ASR, and has meaning ("the sign has changed") only for ASL.

This should be checked against behaviour on an actual CPU or the
Visual 6502 transistor-level simulation of the 6800.