src/simple-a68: Renaming and reformatting
Rename `pnyb` to `prnyb` for consistency across print routines.

Reformat the code to have accumulator name as a separate argument, and
captialized, rather than glomming it on to the instruction name. This lines
things up better, and also makes the accumulator name a bit easier to spot,
which somehow feels like a good thing.