testmc.mc6800: Remove opcodes.pt, moving all tests to opimpl.pt
It was silly having a file full of tests for code mostly in another module,
and made finding the right files to edit confusing. Now all tests are in
opimple.pt. Each opcode test tests both the OPCODES mapping from opcodes.py
and the implementation in opimpl.py, but the latter is the majority of the
code.