testmc.tool.asl: Handle new 6th field produced by cur-142-bld172
Now that we can have five or six fields, we no longer have the ability
to detect the non-escaped-space bug in older (2008 through
asl-current-142-bld151) versions of ASL if the symbol has only one
space in it.

In theory we could restore this (and probably even correctly parse
files from those versions) if we checked which version of ASL was
used, but that's a lot of work for very little benefit. I don't know
of anyone who was distributing those versions of ASL, much less might
still be doing so in supported systems now.