testmc.mc6800: Implement LDAAm, extract test params
We are now starting to genericise our test parameters, in this case between
LDAA (taking a one-byte immediate argument) and LDAAm (taking a two-byte
address whence to load the data). This should work in more or less the same
way for the other six LDA instructions, but we will probably see more
duplication in the test_* functions themselves that we can also extract.