src/cpxtest.a68: Run CPX instructions and store resulting flags
This is to help make sure that the emulator is doing the same thing as
a real CPU. The data in the test are exactly those from cpxparams in
testmc.mc6800.opimpl tests, confirming that the program is doing the
same thing as those simulator unit tests are doing.

Also included is actual output from a JR-200, which is wrong, but seems
more likely to be a problem with this program than a problem with the CPU;
see the test for further details.

The second "test" traces values at a breakpoint, as you could do
manually in a debugger. It was used to debug the code and is kept as
an example of how to do this (until we add this kind of facility to
testmc, anyway). Uncomment the `assert 0` to have pytest print stdout
so you can see the trace.