testmc.mc6800.opimpl: Textual move: PSH/PUL with JSR/RTS etc.
The PSH and PUL instructions do not affect the flags (condition codes); all
other data movement instructions do. For this and other reasons, PSH/PUL
are more closely related, both in functionality and how we test them, to
the other instructions that manipulate the stack (JSR, RTS, etc.). So move
those to be together.