testmc.mc6800.opimpl: Rework LDST test data
We now set up a separate set of test data for instructions doing 16-bit
loads and stores, which makes the test data we're using much more clear.
(Previously we were doing various manipulations to the 8-bit test data in
several different places.)

We also take this chance to improve both the 8- and 16-bit test value sets;
this turned out to be essential to confirming that this change was done
correctly.