testmc.mc6800.opcodes.Instructions: Add subscript access
For our automated tests it's convenient to be able to generate the opcode
name; this makes it easier to look up than using `getattr(I, opname)`.

`Instructions` is not really such a good name for this object becuase,
since the name also includes the addressing mode, they're really opcode
names, not instruction names. But renaming this throughout the project
would have a broad effect, and it feels like using `OP.LDAAx` in a test is
less clear than `I.LDAAx`.