testmc.mc6800.opimpl LDSTaccum(): Parametrize accumulators (A and B)
The input parameter is a template for the opcode name from Instructions.
The output parameters passed to the test are the opcode name itself and a
function to construct a Registers object with the parametrized accumulator
set to a given value, along with whatever other registers and flags the
caller cares to pass in.

Developers less familiar with functional programming, particularly
generating functions to pass back to callers, may find that understanding
this code requires careful study.