testmc.mc6800.opimpl runop(): Better register/flag default values
The previous "set registers and flags to odd values" code was setting
unused flags to 1, which often is H, I and C. Having these as upper- rather
than lower-case letters outside incorrect flags of interest made reading
the incorrect values more difficult, so now we always start runop() with
all flags clear (except those the test explicitly specifies should be set).

We also now have different and more obvious specific values than `not 0`
(== 1) for the A, B and X registers to better detect incorrect overwrites
of them when they should remain untouched. We were already catching an
accidental `0` write, but we now catch accidental `1` writes (`1` being
kind of a common value as well), and also accidental writes of the same
value to more than one of the registers that should remain untouched.