src/bigint: Extract convascdigit tests to a module
This demonstrates how we can re-use the same tests for two different
assembly outputs. The most obvious application is to test the same
routine/API written for two different processors, but it could also be used
to test alternate implementations, etc.

We move the `test_convascdigit_*` tests from `bigint.pt` to `bigint.py`.
We then enable asserting rewriting for that module with
`pytest.register_assert_rewrite()` and import the functions back into
`bigint.pt`, renamed to have `_6502` appended.

The renaming is important, since pytest will describe failures as being at
the location of the function definition (`bigint.py:58`) rather than as
coming from the importing file. However, whether the test is collected and
run depends on the name in the importing file: the library file need not
prefix function names with `test_` so long as they're renamed to have that
prefix in the importing file.

Currently this depends upon the APIs using the same-named registers and
flags with the same semantics. (E.g. C flag is fine when adding, but works
differently when subtracting on 6800 vs. 6502.) We'll have to work out ways
to deal with that.

For this commit versus its parent, the total test count has been verified
to be the same (893) and the individual tests run from bigint.pt have also
had their names compared to confirm they're still all there.