testmc.mc6800.opimpl: Extract test parameters to data structures
This will let us re-use them when we add tests for extended and
indexed addressing modes.

We confirmed that the same number of tests are running now as in
the previous commit.

Additionally, explain in a comment why we manually specify in the
test data the N, Z and C flag values rather than calculating them
automatically.