testmc.mc6800.opimpl: Rework default and generated memory locations
runop() now has a default memory size of $402, and deposits the opcode and
its operands at an arbitrary location between $300 and $3FF, unless
startaddr is specified. (This is moved up from the zero page; using that
made me worried that certain tests might accidentally stomp on their own
code, which has happened in the past.)

v8() and v16() are provided to generate sort-of pseudo-random byte and word
values based on an input value, for use in generating addresses, offsets,
and the like. v() is of course always $00-$FF; v16 is $080-$17F unless a
different offset is specified.

Remove the memsize specification on most tests where it was smaller than
our new default.

test_RTS() was reworked to use the initialization parameters for runop()
rather than generate and tweak its own Machine; presumably it was written
before runop() had these parameters. (We leave the `machine` parameter to
runop() because we may need it later.)

Remove the unused extendargs() which was designed to help do what we're now
doing better here.