testmc.mc6800.opimpl: Remove addresses from test data
LDSTparams and LDST16params included addresses so that we'd be testing at
different addresses; this has now been replaced by v16() and v8(). Also we
were parametrizing with offsets generating further tests just so we could
get different offsets; using v8() on the existing tests means we no longer
need LDSToffsets parametrization and reduces the number of tests a bit.