testmc.mc6800: Do all four A/B/m/X modes in one test for COM, NEG, INC
Doing separate tests for each addressing mode was not terribly useful since
the most common error is a logic error that affects all addressing modes,
which would produce four failures saying essentially the same thing; we now
produce just one if there's any logic issue.

This also speeds the tests, reduces the amount of code, and better exposes
a pattern that we hope to refactor out.