testmc.mc6800: Start to orthogonalize COM/NEG/INC tests
All three tests now specify a full set of arithmetic flags (N, Z, V, C),
assigning None to the flags that should not be changed, and the required
output flags are now set once rather than separately for each of the four
assertions (ACCA, ACCB, extended, indexed).

For INC we also pseudo-randomly (based on the input value) set any ignored
(None) flags for each case so that we can now detect when a flag is updated
when it should not be.