testmc.mc6800: Consolidate CPX addressing modes into one test
We consolidate the immediate, direct (zero page) and extended (memory) for
the same reasons that we did this for the ABmx tests earlier. However, we
do not generalize this yet (either to a general function for these three
modes or by trying to extend runABmx()) because it's not clear if we'll
need this elsewhere.

Generalizing runABmx() to non-ACCx opcode groups looks like it would add
considerable complexity to it; at this point it looks like a bad idea.