testmc.mc6800: Implement all remaining AND instructions
This introduces a new runbinary() function, similar to runABmx() except
that it does the standard eight opcodes (A and B registers × immediate,
direct, extended and indexed addressing modes) for binary operations.

This has almost entirely duplicated code for A and B registers which should
be de-duplicated, and also shares a fair amount of code with runABmx(),
which itself should probably be renamed to rununary().