testmc.mc6800: Refactor and shrink the load/store tests
LDA: use runbinary(). LDA can be considered a "binary" instruction if we
explicitly provide the intial value of the accumulator, which we were doing
anyway to ensure that we weren't getting a false negative. And of course
the expected result is just arg1.

LDX: Use a single test for all addressing modes rather than separate ones
for each addressing mode, for the same reasons as previously done for other
operations.

CLR: Use rununary(), and make sure we have enough `arg` values to get tests
at various memory locations.

STA: Do all three addressing modes in one test, for the usual reason: most
of the code for all three is common.