exec/tmc68/hello: "Hello world" program
This ends with an invalid opcode, causing the simulator to abort with an
exception. It's not clear if there's a better way of exiting (perhaps via
an I/O instruction?), but this will do for a moment. This line can be
commented out to instead go into an infinite loop.