bin/tmc6800: Run a 6800 .p AS output file in the simulator
This is currently quite basic and sets up only character output on $C000.
However, this is enough to run `exe/tmc6800/hello.a68`.

Input is not yet set up; this will require setting up the input tty for raw
mode, if available.

This may fail if the .p file has no entrypoint. Probably we should read the
reset vector for the start point in this case.