testmc.mc6800: LSB() MSB() moved from testmc; these are not generic
Since these depend on whether the architecture is big- or little-endian,
they should be supplied to the external world by the testmc.*.machine
modules, and of course also re-exported by the top level module for that
architecture.

In the mc6800 case we need to actually implement it in the opimpl module
because it's used there, and let the exports filter up.