src/{m65,asl,asxxxx}: CPU and assembler subdirs
All generic MOS 6502 code now goes under src/m65/. However, code specific
to testing assemblers (which may be 6502 code only because that's the
processor we happened to choose for the file that tests assembler features)
goes under asl/ for Macroassembler AS and asxxxx/ for the ASxxxx assembler
suite.

The ASxxxx build system currently builds everything into obj, with no
subdirectory structure; this should probably be changed.