Test,src/asxxxx/: Move ASxxxx object files to $BUILD/obj/asxxxx
While ideally ASxxxx object files should follow a the full directory
structure under src, along with the ASL object files, this requires in the
long run developing some way to distinguish the two types of source files
if we're going to intermix them. That's not likely to be needed any time
soon, so hack this for the moment by just creating a separate asxxxx/
subdir under the object dir and dump all ASxxxx object files in there, so
at least they're not conflicting with other stuff.