src/tmc68/bios: prchar/rdchar BIOS definitions and code for simulator
tmc68/bios.a68 is the definition file for the tmc68 "BIOS" used in the
MC6800 simulator when unit testing and running programs at the Unix command
line. It currently supports `prchar` to print the character in A and
`rdchar` to do a blocking read of a character, returned in A.

The unit test and command-line-run frameworks are expected to separately
load the actual MC6800 code for these routines and ensure that the
simulator is configured to do I/O in the appropriate way (in-memory streams
for unit tests and stdin/stdout for running things at the command line).

tmc68/bioscode is the BIOS implementation itself. It assembles `prchar` and
`rdchar` at the locations defined in `bios.a68` and has some checks to
ensure that the code is at the defined locations and is not overlapping
other BIOS code.

We could just generate the definition file, but it seems better to fix the
locations of the BIOS routines so that they change only when we explicitly
want them to, rather than whenever bioscode.a68 changes.