src/*68/hello,bios: Bring prnl into BIOS and print \n only
The proper thing to print to start a new line varies between platforms and
many platform BIOSes provide a routine to do this. So pull `prnl` out of
the hello program and instead use a BIOS call.

For the tmc68 platform this requires adding `prnl` to the BIOS. In this
case we are currently doing output in cooked mode, so we should print only
a '\n'.

This changes the semantics of newlines enough that it's easiest to rewrite
the hello program to:
- print the newline _before_ the hello message (so the chars we type will
  appear on that same line after the message), and
- echo only the non-command chars, just quietly consuming the Return and
  Ctrl-C that reprint the message and exit, respectively.