convascdigit → qdigit, and extract to its own files
ASCII digit parser renamed to match our new naming convention of `q` prefix
for parsers.

This routine is now used both in the 6502 bigint library and the 6800
proto-monitor, so it's our first "library" routine to split out into its
own file and test separately.

We currently use it by `include`ing it directly in the file of other
routines that use it, even when unit testing, since routines that call
qdigit are unit tested themselves. Obviously this could lead to duplication
in more complex programs; when we reach the point where this happens we'll
have to decide how to handle this. Probably the easiest solution is to use
include guards in the qdigit.aNN file.