src/mc68/cjsmon: Extract from simple-a68 and convert to use BIOS
The start of the machine-language monitor code had been dumped in the
simple test for convenience; now it gets its own file. The internal prchar
etc. routines are dropped and it now uses a BIOS definition, with the unit
tests loading the usual tmc68 BIOS.

prnyb had to be rewritten to use the I/O system; it was previously just
checking a single value written to the test I/O memory location because
that test was written before the I/O framework was available.