src/: Change `M` fixture (a Machine instance) to `m`
The convention is that upper-case `M` would represent a constant or class
(i.e., the Machine class itself) rather than an instance, which by
convention is referenced by a lower-case variable. This also now matches
the convention we use for machine instances in testmc.

We leave `R` upper-case because it is the Registers class.

We also leave `S` upper-case because, though it is an instance of a SymTab,
we use it more like a constant holding our symbol table, and also
lower-case `s` is often used in tests for a string.