Test: Extract astest(); reorganize list of things to build
The astest() function sets up the structure for the unit test framework to
test a "library" file, i.e., one that's tested standalone but designed to
be included in an application or another library file. It automatically
provides `cpu` (based on the file extension) and `org` (which can be
overridden) directives, and includes the `std` definitions file for that
CPU. This greatly reduces the amount of information specified for many of
the individual items to be built.

Also, reorganize the list of items to be built into sections by their role.
This is a bit rough, but is headed in the right direction.