src/mc68/cjsmon: Use regular command calling convention for `Q`
Previously, the quit command depended on the code that found no command to
call then pulling the return to the .parse loop off the stack. We can make
this command not a special case by extracting a command subroutine that
does its own stack manipulation. This makes our command handling more
flexible later on, and I think is also simpler.

This now gives us a location guaranteed to be on the exit path and nowhere
else, so we can use this to simplify the calling procedure of our command
loop tests (which all end with a quit command), and better explain what
we're doing and why, while we're at it.

We also use the new routine's RTS as the handy target of conditional exits.