src/mc68/cjsmon: Replace ~ beep command with ^ print newline
This is better in unit tests because it helps avoid confusing error
indications with expected output (though we may get a bit more
sophisticated about this later) and also this command may be useful later
for separating output as we work out exactly where and how we print
newlines.

`^` is available without shifting on Japanese keyboards; it's to the
left of the `¥` key, which in turn is to the left of backspace.