XXX src/mc68/cjsmon: Replace hardcoded command char parsing with a table
`cmdtbl` is a user-tweakable pointer to a command table of 3-byte
entries containing the command char and the address to call. Command
char $00 terminates the table.

This will eventually be extended to allow continuation entries in the
table so that we can prefix a user-specified table to the default one,
allowing us to add user-specified commands that may override (or
re-use?) existing commands.

During this we've also exposed an issue with when we print newlines. We
start a new line (printed by the input routine) when we hit CR, but after
that we want to print newlines only when necessary: long lines output from
a command should have their own newlines, but several short pieces of
output from commands should be on the same line, and no output should not
generate a new line. Here we see that entering some space (do-nothing)
commands prints nothing on the line, but the command loop adds a newline
anyway before printing the next prompt.

We'll need to work out how to fix this. Probably it involves individual
commands communicating whether they've printed enough that a newline is
necessary after their output, so that the command loop can separate output
from a sequence of such commands on to their own lines, but not terminate
the final output with a newline unless it's "long enough." (This sounds as
if it will print short output and the following prompt on the same line,
which would actually be a good thing.)