src/mc68/cjsmon: Introduce Mcur; hexdump is now a regular command
If we parse a hex word, that now sets Mcur, also known as the `*`
variable, representing the "current location" where commands start
operating. `hexdump` is moved to the `x` command and now always starts
at Mcur.

And yes, we're now mixing upper and lower case here. It's not totally
clear yet how we want to deal with this; we probably want to make the
default command set all lower-case (for ease of typing on mixed-case
machines--the JR200 has no caps lock) and on upper-case-only machines
do a case-insensitive comparison for command letters. This will give
us an extra 26 commands (useful for user commands) on mixed-case
machines.

We can also consider if we want some sort of escape system for upper
case on upper-case only machines, e.g., `\c` to input `C` and
displaying upper case letters in inverse video.