src/mc68/cjsmon: Improve `x` examine command, add init, much cleanup
The `x` examine command now has settings for number of bytes per line to
print (`Mexwidth`) and number of lines to print (`exlines`). The former is
currently set to 16 at startup, and and the latter defaults to 8 lines at
startup but may be changed with an argument to `x`.

A new `init` routine has been added to do the minimal setup necessary to
make `x` work sanely; this then drops into the command line interpreter
loop. Most tests of commands going through the command line parser that may
need this setup are now in an "integration tests" section and the main
tmc68 executable now enters via `init`.

Code refactorings and cleanup:
- Each (uninitialized) monitor/parser storage location is now marked with
  `b` or `W` for its size (byte or word).
- Rename `hexdump` to `examine`.
- Remove/fix some obsolete comments.

Test refactorings and cleanup:
- Extract setinput() to set up command arguments for command unit tests.
- Rename cmdloop() to runmon(); make it call `init` instead of `cmdloop`.