Test: Define `UNIT_TESTING equ 1` when assembling files with astest()
Assembly source files that are "library" code depend on the file including
them to set up any system-specific symbols they need, such as zero page
location assignments. When unit testing there is no "main program" doing
this, so the unit test system itself must supply these definitions.

This is a hack to let us do that. The better way to do this is described in
the comments for astest(), but doing it in Bash would increase the already
terrible complexity of the the Bash code, so we're delaying that until we
rewrite Test to do the heavy lifting in Python.