src/**/{qdigit→qhex}.*: New file for hexadecimal parsing library
The other hex parsing routines currently in the monitor will move into here
as well. qdigit is here even though it's a more general any-base digit
parsing routine because it's as good a place as any for it; library clients
will use ifdefs to include just the parts they want.

I'm not totally sure that this is the right way to go, but it's at least a
step towards more modularity while avoiding a huge proliferation of tiny
files.