src/mc68/qhex qhex: Use temp addr instead of B for max digit count
This is not for optimization (though it does save a trivial four cycles in
the parse loop) but to help free up the B register for other purposes.

We also switch from B to A to pass in the maximum number of digits to
parse (in the case of qhex being used).