src/mc68/qhex: Replace qhex_tmp w/ qhex_stat (bits 6..0 preserved)
Rather than require an entire temporary byte that we destroy, we now put
our exit status in the top bit of a location, preserving the remaining bits
so that the byte can be used for other purposes as well. This increases
code size by 5-6 bytes, but lets clients delay the check to see if a number
was successfully read without having to spend more bytes than that doing
their own tests/loads/stores after every call. (This will save a bit of
work in the monitor.)

It also gains a trivial saving of 3 bytes and 6 cycles in the digit loop.