src/mc68/qhex: Fix broken tests and fix/improve docs
Since at least one char is always checked, and the trailing char is
always checked when parsing less than the maximum number of chars, we
need to set that char explicitly in the test data. The tests were passing
to this point because the byte after the test input data happened to be
initialized to $00.

We also had forgotten to test parsing of all-'0' input. This is especially
important in order to show that leading '0's are significant, not ignored.

The header comment wasn't clear that the maximum number of chars to parse
must be from 1 through 4; this is fixed.

Since we need to set the output to $0000 at the start, and preserving it
would be expensive anyway, we might as well make that part of the API in
case having that $0000 after failure is useful to callers.