src/mc68/qhex: Deposit/examine now have separate default values
We introduce the new concept of a "pending value" (Mpending) that is
whatever hexword value (if any) was most recently read before a command.
If present, deposit and examine use this value, otherwise they use their
own individual default values (whatever was last set for each command).
This allows us to pick a convenient block to dump, make changes in the
middle of that block, and confirm them by re-dumping the block without
having to retype a lot of addresses.