src/mc68/cjsmon: Move command table; improve comments; two JSR→BSR
The command table moves down to be with the commands themselves, which
is a better location from both a source code and object code point of
view. (In particular, it puts the entry point back at the ORG address.)

We make byte vs. word usage in the zero page more clear by simply writing
the second address in a comment, which stands out much better for words and
also will show up in a search for the address of the second byte of a word.
(This trick seen in 8 Bit Show and Tell on YouTube.)