src/mc68/qhex qhex: Remove qhex_stat; it wasn't useful
We don't actually need a memory location to hold the parse success status
(we get it from ACCB anyway), and it turned out that, as with so much
bottom-up design, it wasn't useful to the caller. (The monitor calling
needed to keep the previous value and status when a parse failed, so had to
copy both to its own locations.)

So remove all that extra code and the qhex_stat location. We leave the old
commit present, however, in case we want to rethink this and go back to see
how it was done.