src/mc68/qhex qhex: Remove need for temp storage in RAM
We used to store the (1-byte) max digit count in in memory, but since this
is small (3 bits) we can use some shifts to store it in unused bits of ACCB
instead. This adds 3 bytes of code (2 if the temp is not in the zero page)
over the previous version, but it's better not to require the caller to
co-ordinate the memory usage, especially in systems with minimal RAM.

The new implementation of this now uses the Z flag set to indicate at least
one digit has been successfully parsed, or clear to indicate no digits were
parsed. This is mainly to make setting the result flag more efficient, but
also is a slightly more intuitive API, making the branch after the call BEQ
for success and BNE for failure.