src/mc68/cjsmon: Change prompt from `> ` to `^ `
The `> ` prompt is used by many other programs, such as the built-in
monitor on the JR-200, which can make things quite confusing. `^ ` is not
used as a prompt by by anything else I've ever seen, so this minimizes our
changes of conflict on ports to real systems that have other standard
interactive environments.

It is slightly confusing in the unit tests since `^` is now both part of
the prompt and a command we test, but the latter might be changed or go
away at some point anyway.