src/mc68/cjsmon: Call (`k` for "kall") command
This currently enters the subroutine with undefined registers. Once
we have storage for registers we should load the registers with those
values instead.

There's no integration test for `k` → call in the command table; it
doesn't seem worthwhile since that's so simple and everything but that
is covered by other tests.

The tests are refactored to rename setinput() to command_setup()
because we now don't always set the input when we use it. (It still
sets the input to '\x00' when we don't specify any input, but that's
what the simulator would initialize the memory to anyway.)

As for the command letter, we can't use `c` of course because that's a
hexadecimal digit and will be read as a pending value. Also, `k` is
conveniently adjacent to `j`, which will probably be a "jump/JMP"
command at some point.

To play with this and get an idea of how convenient the separate
current addresses per-command is, you can use the following command to
enter the routine `LDA #00; STA $500E; RTS`, call it, and view the
routine and what it deposited:

    5000:86 00 b7 500e 39: 5001:aa: 5000k 5000x2

Then continue with `:NN:kx` sequences, replacing `NN` with any hex
digits you like, to watch them appear at $500E due to that call.